refactor: make run the primary schema target
